:sparkles: Added Example Input Download, Refactor
You can now use -s flag to download example input from the problem (as long as that problem is
