+++ src/aoc_util/__init__.py
@@ -1,4 +1,5 @@
 from rich.traceback import install
-from .helper import read
+
+from .helper import read  # noqa
 
 install(show_locals=True)
