+++ src/aoc_util/readme.py
@@ -1,9 +1,10 @@
 """Used to automatically update README to reflect stars earned"""
 
-from typing import Optional
 from datetime import date
-from bs4 import BeautifulSoup
+from typing import Optional
+
 import pandas as pd
+from bs4 import BeautifulSoup
 
 from aoc_util.aoc_requests import get_aoc_page
 
@@ -56,7 +57,7 @@ def update_readme(new_table: Optional[str] = None) -> None:
     if new_table is None:
         new_table = get_aoc_stars()
     # GET current README file
-    with open("README.md", "r") as f:
+    with open("README.md") as f:
         md = f.read()
     # remove current table (Only Element under AOC Star Summary)
     new = md.split("#")
